✨ Add param `resolve_last` to call smart contract public method from specific address
