+++ pubspec.yaml
@@ -3,7 +3,7 @@ name: archethic_lib_dart
 description: Archethic dart library for Flutter for Node and Browser. This library aims to provide a easy way to create Archethic transaction and to send them over the network
 homepage: https://github.com/archethic-foundation/libdart
 
-version: 3.3.16
+version: 3.3.17
 
 environment:
   sdk: ">=3.3.0 <4.0.0"
